+++ sismic/__init__.py
@@ -1,5 +1,5 @@
 __description__ = 'Sismic Interactive State Machine Interpreter and Checker'
-__version__ = '0.20.0'
+__version__ = '0.20.1'
 __url__ = 'https://github.com/AlexandreDecan/sismic/'
 __author__ = 'Alexandre Decan'
 __email__ = 'alexandre.decan@lexpage.net'